[fix] cleanup/collapse/reconcile oss and ee
