+++ api/ee/src/core/organizations/exceptions.py
@@ -1,10 +1,6 @@
-"""Domain-level exceptions for organizations."""
+"""Domain-level exceptions for organizations (EE-specific)."""
 
-from oss.src.core.organizations.exceptions import (  # noqa: F401 — shared base/types
-    OrganizationError,
-    OrganizationCreationNotAllowedError,
-    LastOrganizationError,
-)
+from oss.src.core.organizations.exceptions import OrganizationError
 
 
 class OrganizationSlugConflictError(OrganizationError):
@@ -13,13 +9,4 @@ class OrganizationSlugConflictError(OrganizationError):
     def __init__(self, slug: str, message: str = None):
         self.slug = slug
         self.message = message or f"Organization slug '{slug}' is already in use."
-        super().__init__(self.message)
-
-
-class OrganizationNotFoundError(OrganizationError):
-    """Raised when an organization is not found."""
-
-    def __init__(self, organization_id: str, message: str = None):
-        self.organization_id = organization_id
-        self.message = message or f"Organization with id '{organization_id}' not found."
         super().__init__(self.message)
