--- api/ee/src/models/api/workspace_models.py
@@ -1,71 +1,0 @@
-from datetime import datetime
-from typing import Optional, List, Dict
-
-from pydantic import BaseModel
-
-from ee.src.models.api.api_models import TimestampModel
-from ee.src.core.access.permissions.types import Permission
-
-
-class WorkspacePermission(BaseModel):
-    # Role slugs are dynamic (env-overridable via AGENTA_ACCESS_ROLES);
-    # validation against the effective scope catalog happens at the API
-    # boundary via `ee.src.core.access.controls.get_role`.
-    role_name: str
-    role_description: Optional[str] = None
-    permissions: Optional[List[Permission]] = None
-
-
-class WorkspaceMember(BaseModel):
-    user_id: str
-    roles: List[WorkspacePermission]
-
-
-class WorkspaceMemberResponse(BaseModel):
-    user: Dict
-    roles: List[WorkspacePermission]
-
-
-class Workspace(BaseModel):
-    id: Optional[str] = None
-
-    name: Optional[str] = None
-    description: Optional[str] = None
-
-    type: Optional[str]
-
-    members: Optional[List[WorkspaceMember]] = None
-
-
-class WorkspaceResponse(TimestampModel):
-    id: str
-
-    name: Optional[str] = None
-    description: Optional[str] = None
-
-    type: Optional[str]
-
-    organization: str
-
-    members: Optional[List[WorkspaceMemberResponse]] = None
-
-
-class CreateWorkspace(BaseModel):
-    name: Optional[str] = None
-    description: Optional[str] = None
-
-    type: Optional[str] = None
-
-
-class UserRole(BaseModel):
-    email: str
-    role: Optional[str] = None
-
-    organization_id: str
-
-
-class UpdateWorkspace(BaseModel):
-    name: Optional[str] = None
-    description: Optional[str] = None
-
-    updated_at: Optional[datetime] = None
