--- api/ee/src/models/db_models.py
@@ -1,13 +1,0 @@
-# Membership models moved to OSS (shared schema); re-exported here so existing
-# EE imports keep working.
-from oss.src.models.db_models import (
-    OrganizationMemberDB,
-    WorkspaceMemberDB,
-    ProjectMemberDB,
-)
-
-__all__ = [
-    "OrganizationMemberDB",
-    "WorkspaceMemberDB",
-    "ProjectMemberDB",
-]
