+++ api/oss/src/models/api/workspace_models.py
@@ -1,9 +1,10 @@
 from datetime import datetime
-from typing import Optional, List
+from typing import Any, Dict, Optional, List
 
 from pydantic import BaseModel
 
 from oss.src.models.api.api_models import TimestampModel
+from oss.src.core.access.permissions.types import Permission
 
 
 class Workspace(BaseModel):
@@ -11,6 +12,24 @@ class Workspace(BaseModel):
     name: str
     description: Optional[str] = None
     type: Optional[str]
+
+
+class WorkspacePermission(BaseModel):
+    # Role slugs are dynamic (env-overridable via AGENTA_ACCESS_ROLES in EE);
+    # validation against the effective scope catalog happens at the API boundary.
+    role_name: str
+    role_description: Optional[str] = None
+    permissions: Optional[List[Permission]] = None
+
+
+class WorkspaceMember(BaseModel):
+    user_id: str
+    roles: List[WorkspacePermission]
+
+
+class WorkspaceMemberResponse(BaseModel):
+    user: Dict[str, Any]
+    roles: List[WorkspacePermission]
 
 
 class InviteRequest(BaseModel):
@@ -43,6 +62,7 @@ class WorkspaceResponse(TimestampModel):
     description: Optional[str] = None
     type: Optional[str]
     organization: str
+    members: Optional[List[WorkspaceMemberResponse]] = None
 
 
 class CreateWorkspace(BaseModel):
