+++ api/oss/src/routers/organization_router.py
@@ -11,9 +11,8 @@ from oss.src.utils.common import APIRouter
 from oss.src.models.api.organization_models import (
     Organization,
     OrganizationDetails,
-    OrganizationMember,
     CreateOrganizationPayload,
-    UpdateOrganizationPayload,
+    OrganizationUpdate,
 )
 from oss.src.services import organization_service
 from oss.src.services.organization_service import (
@@ -26,30 +25,34 @@ from oss.src.models.api.workspace_models import (
     InviteRequest,
     ResendInviteRequest,
     InviteToken,
+    CreateWorkspace,
+    UpdateWorkspace,
+    WorkspaceResponse,
 )
-from oss.src.core.access.permissions.service import check_action_access
+from oss.src.core.access.permissions.service import (
+    check_action_access,
+    check_rbac_permission,
+)
 from oss.src.core.access.permissions.types import Permission
+from oss.src.core.organizations.exceptions import (
+    OrganizationCreationNotAllowedError,
+)
 from oss.src.services.db_manager import get_user_org_and_workspace_id
 
 
-def _role_description(role: str) -> str:
-    """Resolve a workspace-role description from the effective access-controls
-    catalog (env-overridable via AGENTA_ACCESS_ROLES in EE; OSS uses defaults)."""
-    from oss.src.core.access.controls import get_role_description
-
-    return get_role_description("workspace", role) or ""
-
-
 if is_ee():
-    from ee.src.services import db_manager_ee
-
     from ee.src.core.access.entitlements.service import (
         check_entitlements,
         scope_from,
         Tracker,
         Gauge,
+        Flag,
         NOT_ENTITLED_RESPONSE,
     )
+    from ee.src.core.organizations.service import (
+        update_organization as update_organization_ee,
+    )
+    from ee.src.core.organizations.exceptions import OrganizationSlugConflictError
 
 router = APIRouter()
 
@@ -74,7 +77,7 @@ async def list_organizations(
         user_org_workspace_data: dict = await get_user_org_and_workspace_id(
             request.state.user_id
         )
-        organizations_db = await db_manager_ee.get_organizations_by_list_ids(
+        organizations_db = await db_manager.get_organizations_by_list_ids(
             user_org_workspace_data["organization_ids"]
         )
         workspaces_by_org = {}
@@ -129,94 +132,15 @@ async def fetch_organization_details(
             content={"detail": "Not a member of this organization."},
         )
 
-    organization_workspaces = [
-        workspace_db
-        for workspace_db in await db_manager.get_workspaces()
-        if str(workspace_db.organization_id) == str(organization_id)
-    ]
-    active_workspace = next(iter(organization_workspaces), None)
-    if not active_workspace:
+    # Both editions: return the org with its default workspace and real project
+    # members + permissions (demo orgs skip members). tags/meta are included.
+    organization = await db_manager.get_organization_by_id(
+        organization_id=organization_id
+    )
+    if organization is None:
         return {}
 
-    organization_owner = await db_manager.get_organization_owner(
-        organization_id=organization_id
-    )
-
-    default_project = await db_manager.get_default_project_by_organization_id(
-        organization_id=organization_id
-    )
-    project_invitations = (
-        await db_manager.get_project_invitations(project_id=str(default_project.id))
-        if default_project
-        else []
-    )
-
-    organization_db = await db_manager.get_organization_by_id(
-        organization_id=organization_id
-    )
-
-    invited_members = [
-        {
-            "user": {
-                "id": str(invitation.id),
-                "email": invitation.email,
-                "username": invitation.email.split("@")[0],
-                "status": "pending" if not invitation.used else "member",
-                "created_at": str(invitation.created_at),
-            },
-            "roles": [
-                {
-                    "role_name": invitation.role or "viewer",
-                    "role_description": _role_description(invitation.role or "viewer"),
-                }
-            ],
-        }
-        for invitation in project_invitations
-    ]
-
-    owner = [
-        {
-            "user": OrganizationMember(
-                id=str(organization_owner.id),
-                email=str(organization_owner.email),
-                status="member",
-                username=str(organization_owner.username),
-                created_at=str(organization_owner.created_at),
-            ).model_dump(exclude_none=True),
-            "roles": [
-                {
-                    "role_name": "owner",
-                    "role_description": "Can fully manage the workspace, including adding and removing members.",
-                }
-            ],
-        }
-    ]
-
-    # Merge invited members
-    members = invited_members + owner
-
-    return OrganizationDetails(
-        id=str(organization_db.id),
-        slug=str(organization_db.slug),
-        #
-        name=str(organization_db.name),
-        description=str(organization_db.description),
-        #
-        flags=organization_db.flags,
-        tags=organization_db.tags,
-        meta=organization_db.meta,
-        #
-        owner_id=organization_db.owner_id,
-        #
-        default_workspace={
-            "id": str(active_workspace.id),
-            "name": str(active_workspace.name),
-            "description": str(active_workspace.description),
-            "type": active_workspace.type,  # type: ignore
-            "members": members,
-        },
-        workspaces=[str(workspace.id) for workspace in organization_workspaces],
-    ).model_dump(exclude_unset=True)
+    return await db_manager.get_org_details(organization)
 
 
 @router.post(
@@ -425,195 +349,312 @@ async def _check_org_owner(organization_id: str, user_id: str) -> bool:
     return organization is not None and str(organization.owner_id) == str(user_id)
 
 
-if not is_ee():
-    # In EE these endpoints come from the EE organization router (with RBAC
-    # and entitlement gates); registering them here too would shadow it.
-    from fastapi import HTTPException
-
-    from oss.src.services.commoners import create_organization_for_user
-    from oss.src.core.organizations.exceptions import (
-        OrganizationCreationNotAllowedError,
-    )
-
-    @router.post("/", operation_id="create_organization")
-    async def create_organization(
-        request: Request,
-        payload: CreateOrganizationPayload,
-    ):
-        """Create a new organization."""
-        try:
-            user = await db_manager.get_user(request.state.user_id)
-            if not user:
-                return JSONResponse(
-                    {"detail": "User not found"},
-                    status_code=404,
-                )
-
-            organization = await create_organization_for_user(
-                user_id=UUID(str(user.id)),
-                organization_name=payload.name,
-                organization_description=payload.description,
+@router.post("/", operation_id="create_organization")
+async def create_organization(
+    request: Request,
+    payload: CreateOrganizationPayload,
+):
+    """Create a new organization."""
+    try:
+        user = await db_manager.get_user(request.state.user_id)
+        if not user:
+            return JSONResponse(
+                {"detail": "User not found"},
+                status_code=404,
             )
 
+        # OSS-owned; EE subscription provisioning layers in via is_ee() inside.
+        from oss.src.services.commoners import (  # noqa: PLC0415
+            create_organization_for_user,
+        )
+
+        organization = await create_organization_for_user(
+            user_id=UUID(str(user.id)),
+            organization_name=payload.name,
+            organization_description=payload.description,
+        )
+
+        return JSONResponse(
+            {
+                "id": str(organization.id),
+                "name": organization.name,
+                "description": organization.description,
+            },
+            status_code=201,
+        )
+
+    except OrganizationCreationNotAllowedError as e:
+        raise HTTPException(
+            status_code=403,
+            detail=e.message,
+        ) from e
+
+    except Exception:
+        log.error(
+            "Unexpected error while creating organization",
+            exc_info=True,
+        )
+        raise HTTPException(
+            status_code=500,
+            detail="An internal error occurred while creating the organization.",
+        )
+
+
+@router.put("/{organization_id}", operation_id="update_organization")
+@router.patch("/{organization_id}", operation_id="patch_organization")
+async def update_organization(
+    request: Request,
+    organization_id: str,
+    payload: OrganizationUpdate,
+):
+    if (
+        not payload.name
+        and not payload.description
+        and not payload.slug
+        and not payload.flags
+    ):
+        return JSONResponse(
+            {"detail": "Please provide a field to update"},
+            status_code=400,
+        )
+
+    if not await _check_org_owner(organization_id, request.state.user_id):
+        return JSONResponse(
+            {"detail": "You do not have permission to perform this action"},
+            status_code=403,
+        )
+
+    try:
+        if is_ee():
+            # EE accepts slug/flags; flag changes are entitlement-gated and run
+            # through SSO/domain/auth-lockout validation in the EE service.
+            if payload.flags is not None:
+                check, _, _ = await check_entitlements(
+                    key=Flag.ACCESS,
+                    scope=scope_from(organization_id=UUID(organization_id)),
+                )
+                if not check:
+                    return NOT_ENTITLED_RESPONSE(Tracker.FLAGS)
+
+            organization = await update_organization_ee(organization_id, payload)
+            return organization
+
+        organization = await db_manager.update_organization(
+            organization_id=organization_id,
+            values_to_update=payload.model_dump(exclude_unset=True),
+        )
+        return JSONResponse(
+            {
+                "id": str(organization.id),
+                "name": organization.name,
+                "description": organization.description,
+            },
+            status_code=200,
+        )
+
+    except ValueError:
+        # Slug validation errors (format, immutability, personal org, etc.)
+        return JSONResponse(
+            {"detail": "Invalid request data for organization update."},
+            status_code=400,
+        )
+    except Exception as exc:
+        if is_ee() and isinstance(exc, OrganizationSlugConflictError):
             return JSONResponse(
                 {
-                    "id": str(organization.id),
-                    "name": organization.name,
-                    "description": organization.description,
+                    "detail": "Slug already in use. Please select another slug or contact your administrator."
                 },
-                status_code=201,
+                status_code=409,
             )
+        log.error(
+            "Unexpected error while updating organization",
+            exc_info=True,
+        )
+        raise HTTPException(
+            status_code=500,
+            detail="An internal error occurred while updating the organization.",
+        )
 
-        except OrganizationCreationNotAllowedError as e:
-            raise HTTPException(
-                status_code=403,
-                detail=e.message,
-            ) from e
 
-        except Exception:
-            log.error(
-                "Unexpected error while creating organization",
-                exc_info=True,
-            )
-            raise HTTPException(
-                status_code=500,
-                detail="An internal error occurred while creating the organization.",
-            )
+@router.post(
+    "/{organization_id}/transfer/{new_owner_id}",
+    operation_id="transfer_organization_ownership",
+)
+async def transfer_organization_ownership(
+    request: Request,
+    organization_id: str,
+    new_owner_id: str,
+):
+    """Transfer organization ownership to another member."""
+    if not await _check_org_owner(organization_id, request.state.user_id):
+        return JSONResponse(
+            {"detail": "Only the organization owner can transfer ownership"},
+            status_code=403,
+        )
 
-    @router.put("/{organization_id}", operation_id="update_organization")
-    @router.patch("/{organization_id}", operation_id="patch_organization")
-    async def update_organization(
-        request: Request,
-        organization_id: str,
-        payload: UpdateOrganizationPayload,
-    ):
-        if not payload.name and not payload.description:
+    try:
+        organization = await db_manager.transfer_organization_ownership(
+            organization_id=organization_id,
+            new_owner_id=new_owner_id,
+            current_user_id=str(request.state.user_id),
+        )
+
+        return JSONResponse(
+            {
+                "organization_id": str(organization.id),
+                "owner_id": str(organization.owner_id),
+            },
+            status_code=200,
+        )
+
+    except ValueError:
+        log.warning(
+            "Invalid organization ownership transfer request",
+            exc_info=True,
+        )
+        return JSONResponse(
+            {"detail": "Invalid organization or new owner for ownership transfer"},
+        )
+    except Exception:
+        log.error(
+            "Unexpected error while transferring organization ownership",
+            exc_info=True,
+        )
+        raise HTTPException(
+            status_code=500,
+            detail="An internal error occurred while transferring organization ownership.",
+        )
+
+
+@router.delete("/{organization_id}", operation_id="delete_organization")
+async def delete_organization(
+    request: Request,
+    organization_id: str,
+):
+    """Delete an organization (owner only)."""
+    if not await _check_org_owner(organization_id, request.state.user_id):
+        return JSONResponse(
+            {"detail": "You do not have permission to perform this action"},
+            status_code=403,
+        )
+
+    try:
+        org_count = await db_manager.count_organizations_by_owner(
+            str(request.state.user_id)
+        )
+        if org_count <= 1:
             return JSONResponse(
-                {"detail": "Please provide a field to update"},
+                {
+                    "detail": "Cannot delete your last organization. You must have at least one organization."
+                },
                 status_code=400,
             )
 
+        await db_manager.delete_organization(organization_id)
+
+        log.info(
+            "[organization] organization deleted",
+            organization_id=organization_id,
+            user_id=request.state.user_id,
+        )
+
+        return JSONResponse(
+            {"detail": "Organization deleted successfully"},
+            status_code=200,
+        )
+
+    except Exception:
+        log.error(
+            "Unexpected error while deleting organization",
+            exc_info=True,
+        )
+        raise HTTPException(
+            status_code=500,
+            detail="An internal error occurred while deleting the organization.",
+        )
+
+
+@router.post(
+    "/{organization_id}/workspaces/",
+    operation_id="create_workspace",
+    response_model=WorkspaceResponse,
+)
+async def create_workspace(
+    request: Request,
+    organization_id: str,
+    payload: CreateWorkspace,
+) -> WorkspaceResponse:
+    """Create a new workspace in an organization (owner only)."""
+    try:
         if not await _check_org_owner(organization_id, request.state.user_id):
             return JSONResponse(
                 {"detail": "You do not have permission to perform this action"},
                 status_code=403,
             )
 
-        try:
-            organization = await db_manager.update_organization(
-                organization_id=organization_id,
-                values_to_update=payload.model_dump(exclude_unset=True),
-            )
+        if not payload.name:
             return JSONResponse(
-                {
-                    "id": str(organization.id),
-                    "name": organization.name,
-                    "description": organization.description,
-                },
-                status_code=200,
-            )
-        except Exception:
-            log.error(
-                "Unexpected error while updating organization",
-                exc_info=True,
-            )
-            raise HTTPException(
-                status_code=500,
-                detail="An internal error occurred while updating the organization.",
+                {"detail": "Please provide a name to create a workspace"},
+                status_code=400,
             )
 
-    @router.post(
-        "/{organization_id}/transfer/{new_owner_id}",
-        operation_id="transfer_organization_ownership",
-    )
-    async def transfer_organization_ownership(
-        request: Request,
-        organization_id: str,
-        new_owner_id: str,
-    ):
-        """Transfer organization ownership to another member."""
-        if not await _check_org_owner(organization_id, request.state.user_id):
+        return await organization_service.create_new_workspace(
+            payload, organization_id, request.state.user_id
+        )
+
+    except Exception:
+        log.error(
+            "Unexpected error while creating workspace",
+            exc_info=True,
+        )
+        raise HTTPException(
+            status_code=500,
+            detail="An internal error occurred while creating the workspace.",
+        )
+
+
+@router.put(
+    "/{organization_id}/workspaces/{workspace_id}",
+    operation_id="update_workspace",
+    response_model=WorkspaceResponse,
+)
+async def update_workspace(
+    request: Request,
+    organization_id: str,
+    workspace_id: str,
+    payload: UpdateWorkspace,
+) -> WorkspaceResponse:
+    """Update a workspace's details (requires EDIT_WORKSPACE permission)."""
+    try:
+        user_org_workspace_data = await get_user_org_and_workspace_id(
+            request.state.user_id
+        )
+        project = await db_manager.get_project_by_workspace(workspace_id)
+        has_permission = await check_rbac_permission(
+            user_org_workspace_data=user_org_workspace_data,
+            project_id=str(project.id),
+            permission=Permission.EDIT_WORKSPACE,
+        )
+        if not has_permission:
             return JSONResponse(
-                {"detail": "Only the organization owner can transfer ownership"},
+                {"detail": "You do not have permission to update this workspace"},
                 status_code=403,
             )
 
-        try:
-            organization = await db_manager.transfer_organization_ownership(
-                organization_id=organization_id,
-                new_owner_id=new_owner_id,
-                current_user_id=str(request.state.user_id),
-            )
-
+        if not payload.name and not payload.description:
             return JSONResponse(
-                {
-                    "organization_id": str(organization.id),
-                    "owner_id": str(organization.owner_id),
-                },
-                status_code=200,
+                {"detail": "Please provide a name or description to update"},
+                status_code=400,
             )
 
-        except ValueError:
-            log.warning(
-                "Invalid organization ownership transfer request",
-                exc_info=True,
-            )
-            return JSONResponse(
-                {"detail": "Invalid organization or new owner for ownership transfer"},
-            )
-        except Exception:
-            log.error(
-                "Unexpected error while transferring organization ownership",
-                exc_info=True,
-            )
-            raise HTTPException(
-                status_code=500,
-                detail="An internal error occurred while transferring organization ownership.",
-            )
+        return await organization_service.update_workspace(payload, workspace_id)
 
-    @router.delete("/{organization_id}", operation_id="delete_organization")
-    async def delete_organization(
-        request: Request,
-        organization_id: str,
-    ):
-        """Delete an organization (owner only)."""
-        if not await _check_org_owner(organization_id, request.state.user_id):
-            return JSONResponse(
-                {"detail": "You do not have permission to perform this action"},
-                status_code=403,
-            )
-
-        try:
-            org_count = await db_manager.count_organizations_by_owner(
-                str(request.state.user_id)
-            )
-            if org_count <= 1:
-                return JSONResponse(
-                    {
-                        "detail": "Cannot delete your last organization. You must have at least one organization."
-                    },
-                    status_code=400,
-                )
-
-            await db_manager.delete_organization(organization_id)
-
-            log.info(
-                "[organization] organization deleted",
-                organization_id=organization_id,
-                user_id=request.state.user_id,
-            )
-
-            return JSONResponse(
-                {"detail": "Organization deleted successfully"},
-                status_code=200,
-            )
-
-        except Exception:
-            log.error(
-                "Unexpected error while deleting organization",
-                exc_info=True,
-            )
-            raise HTTPException(
-                status_code=500,
-                detail="An internal error occurred while deleting the organization.",
-            )
+    except Exception:
+        log.error(
+            "Unexpected error while updating workspace",
+            exc_info=True,
+        )
+        raise HTTPException(
+            status_code=500,
+            detail="An internal error occurred while updating the workspace.",
+        )
